fix casting errors and include issue

+++ src/Util/Position.cpp
@@ -7,12 +7,14 @@ PTSDPosition PTSDPosition::operator+(const glm::vec2 vec2) const {
 }
 
 PTSDPosition PTSDPosition::FromSDL(int sdlx, int sdly) {
-    return PTSDPosition{static_cast<float>(sdlx) - WINDOW_WIDTH / 2,
-                        static_cast<float>(sdly) - WINDOW_HEIGHT / 2};
+    return PTSDPosition{
+        static_cast<float>(sdlx) - static_cast<float>(WINDOW_WIDTH) / 2.0F,
+        static_cast<float>(sdly) - static_cast<float>(WINDOW_HEIGHT) / 2.0F};
 }
 
 SDLPosition PTSDPosition::ToSDLPosition() const {
-    return SDLPosition(static_cast<int>(this->x + WINDOW_WIDTH / 2),
-                       static_cast<int>(this->y + WINDOW_HEIGHT / 2));
+    return SDLPosition(
+        static_cast<int>(this->x + static_cast<float>(WINDOW_WIDTH) / 2.0F),
+        static_cast<int>(this->y + static_cast<float>(WINDOW_HEIGHT) / 2.0F));
 }
 } // namespace Util
